test: Delete the user that was created after all the tests run
It's annoying to have to do this manually after each test run. Eventually, there will be UI to do this.
