+++ cypress/e2e/happy-path.cy.ts
@@ -1,6 +1,10 @@
 const INDEX = 'http://localhost:5173/';
 
 describe('Loggr End-to-end happy path tests', () => {
+  after(() => {
+    cy.deleteUser();
+  });
+
   it('Loads the page', () => {
     cy.visit(INDEX);
   });
